pyproject.toml
[tool.poetry]
name = "litbee"
version = "0.1.0"
description = "align texts via streamlit "
authors = ["ffreemt"]
license = "MIT"
readme = "README.md"
repository = "https://github.com/ffreemt/litbee"

[tool.poetry.dependencies]
python = "^3.8.3"
logzero = "^1.7.0"
icecream = "^2.1.1"
install = "^1.3.5"
set-loglevel = "^0.1.2"
streamlit-multipage = "^0.0.18"
cchardet = "^2.1.7"
streamlit-aggrid = "^0.2.3"
ezbee = "0.1.0a3"

[tool.poe.executor]
type = "poetry"

[tool.poe.tasks]
memo = "echo poe test or poetry run poe test"
test = "pytest tests"
build = "poetry build"
_publish = "poetry publish"
release = ["test", "build", "_publish"]
lint = { cmd = "pylint litbee" }
format = "black tests litbee"
docstyle = "pydocstyle --convention=google tests litbee"
tunnel = {cmd = "ssh -CN ip_or_hostname_defined_in_hosts -L 9091:127.0.0.1:9091"}

[tool.poetry.dev-dependencies]
pytest = "^7.1"
flake8 = "^3.9.2"
pydocstyle = "^6.1.1"
toml = "^0.10.2"
# tbump = "^6.3.2"
poethepoet = "^0.10.0"

[tool.poetry.scripts]
"litbee" = "litbee.__main__:app"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
minversion = "6.0"
addopts = "-ra -q --doctest-modules"
log_cli = true
testpaths = [
    "tests",
    # "integration",
]

[tool.black]
skip-string-normalization = 0

[tool.isort]
profile = "black"
multi_line_output = 3
